renamed package, added client package and constant class
