+++ .gitignore
@@ -1,0 +1,3 @@
+/MongoDBWeb/nbproject/private/
+/MongoDBWeb/dist/
+/MongoDBWeb/build/
