+++ MongoDBWeb/nbproject/private/private.properties
@@ -1,7 +1,7 @@
-deploy.ant.properties.file=/Users/taiwenjin/Library/Application Support/NetBeans/8.2/config/GlassFishEE6/Properties/gfv3315458821.properties
+deploy.ant.properties.file=C:\\Users\\berth\\AppData\\Roaming\\NetBeans\\8.2\\config\\GlassFishEE6\\Properties\\gfv3839491542.properties
 j2ee.platform.is.jsr109=true
-j2ee.server.domain=/Applications/NetBeans/glassfish-4.1.1/glassfish/domains/domain1
-j2ee.server.home=/Applications/NetBeans/glassfish-4.1.1/glassfish
-j2ee.server.instance=[/Applications/NetBeans/glassfish-4.1.1/glassfish:/Applications/NetBeans/glassfish-4.1.1/glassfish/domains/domain1]deployer:gfv3ee6wc:localhost:4848
-j2ee.server.middleware=/Applications/NetBeans/glassfish-4.1.1
-user.properties.file=/Users/taiwenjin/Library/Application Support/NetBeans/8.2/build.properties
+j2ee.server.domain=C:/Users/berth/AppData/Roaming/NetBeans/8.2/config/GF_4.1.1/domain1
+j2ee.server.home=C:/Program Files/glassfish-4.1.1/glassfish
+j2ee.server.instance=[C:\\Program Files\\glassfish-4.1.1\\glassfish;C:\\Program Files\\glassfish-4.1.1\\glassfish\\domains\\domain1]deployer:gfv3ee6wc:localhost:4848
+j2ee.server.middleware=C:/Program Files/glassfish-4.1.1
+user.properties.file=C:\\Users\\berth\\AppData\\Roaming\\NetBeans\\8.2\\build.properties
