+++ MongoDBWeb/src/java/com/savelives/clients/CrimeClient.java
@@ -1,0 +1,13 @@
+/*
+ * Created by Cheick Berthe on 2017.04.11  * 
+ * Copyright © 2017 Cheick Berthe. All rights reserved. * 
+ */
+package com.savelives.clients;
+
+/**
+ *
+ * @author berth
+ */
+public class CrimeClient {
+    
+}
