+++ MongoDBWeb/src/java/com/savelives/managers/Constants.java
@@ -1,0 +1,15 @@
+/*
+ * Created by Cheick Berthe on 2017.04.12  * 
+ * Copyright © 2017 Cheick Berthe. All rights reserved. * 
+ */
+package com.savelives.managers;
+
+/**
+ *
+ * @author berth
+ */
+final class Constants {
+    public static final String DATABASE_NAME = "SaveLives";
+    public static final String DATABASE_HOSTNAME = "localhost";
+    public static final int DATABASE_PORT = 27017;
+}
