+++ MongoDBWeb/src/java/com/savelives/managers/CrimeCaseController.java
@@ -2,11 +2,9 @@
  * Created by Taiwen Jin on 2017.04.10  * 
  * Copyright © 2017 Taiwen Jin. All rights reserved. * 
  */
-package com.taiwenjin.mongodb;
+package com.savelives.managers;
 
 import com.mongodb.BasicDBObject;
-import com.mongodb.DB;
-import com.mongodb.DBCollection;
 import com.mongodb.MongoClient;
 import com.mongodb.MongoCredential;
 import com.mongodb.ServerAddress;
@@ -34,15 +32,15 @@ public class CrimeCaseController implements Serializable {
     private final String crimeCaseJSONDataURL = "https://data.baltimorecity.gov/resource/4ih5-d5d5.json";
 
     private MongoClient mongoClient;
-    private DB db;
-    private DBCollection coll;
+    private MongoDatabase db;
+    private MongoCollection coll;
 
     public void dataQuery() {
 
         JSONArray jsonArray;
         String password = "CSD@s17";
-        List<ServerAddress> seeds = new ArrayList<ServerAddress>();
-        List<MongoCredential> credentialsList = new ArrayList<MongoCredential>();
+        List<ServerAddress> seeds = new ArrayList<>();
+        List<MongoCredential> credentialsList = new ArrayList<>();
         
         ServerAddress serverAddress = new ServerAddress("localhost", 27017);
         seeds.add(serverAddress);
@@ -52,7 +50,7 @@ public class CrimeCaseController implements Serializable {
         
         //  mongoClient = new MongoClient(seeds, credentialsList);
         mongoClient = new MongoClient(seeds);
-        db = mongoClient.getDB("TestDatabase");
+        db = mongoClient.getDatabase("SaveLives");
         coll = db.getCollection("CrimeCaseCollection");
 
         try {
@@ -85,7 +83,7 @@ public class CrimeCaseController implements Serializable {
                         .append("district", district)
                         .append("coorX", coorX)
                         .append("coorY", coorY);
-                coll.insert(doc);
+                coll.insertOne(doc);
             }
 
         } catch (Exception ex) {
